LogSearch: Improve search speed by early rejecting non-matching files
I'm assuming that a large amount of time spent in search is due to
individually running a regex over each line of each file.

re2's engine should be very adept at scanning large inputs for a given pattern,
but we're currently not giving it the opportunity to do so. Overheads of per-line checking
add up fast.

To speed things up, we first scan the entire file for any instance of our search term.
This may cause false positive matches (eg. if we're searching for a nick but find the
term in a message instead), but no false negatives (if it's there, we'll find it).

Only if the file has proven to _maybe_ contain our match, do we then split it into lines
and scan it properly. In most cases, the repeated work will only be over a single file,
which should not be noticable.