LogSearch: Fix a bug where the calling command is not ignored
We are correctly removing the last line when `reverse and includeToday and filename == today`,
but due to changing from readlines() to split("\n"), the trailing newline of the file
becomes a blank string in the list of lines. This is what gets removed.

We fix this by removing any trailing newline before splitting.