Comic: Restructure to split semantic generation from rendering
We put all the actual message gathering and generation into its own function,
and the process of converting the resulting info to an image in another function.

Besides being a nice seperation of concerns, this is a stepping stone torwards
being able to accept and render customized comics, or save a comic's info
(which is small) instead of its rendered image (which is large) then render it again later.