[URLFollow] Fix AttributeError in Imgur module
For some reason, the imgur module is trying to follow direct .jpg image links and throwing an AttributeError 'NoneType' object has no attribute 'content' as it does so.

For SOME reason, this can cause tells to silently get lost in the ether, never to be seen again? I don't know what's going on.